+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.6 21 August 2026
 1. Import ModuleCaller from psiutils.module_caller
+2. Add edit module with params: mode to edit, text to select (optional)
 
 ## Version 0.0.5 21 August 2026
 1. Rename utilities_bbochat to utilities